+++ back/app/api/item/item_detail/serializer_item_detail.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+from app.models import Item
+
+class ItemSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Item
+        fields = '__all__'
