+++ back/app/api/item/search/serializer_search.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+from app.models import Item
+
+class ItemSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Item
+        fields = '__all__'
