+++ back/app/api/item/search/views_search.py
@@ -5,18 +5,15 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.serializers import Serializer
 
-# from ..middleware import validation_token
-from ..serializers import *
-from ..models import *
-
+from app.models import Item
 from app.middleware import validation_token
+from .serializer_search import ItemSerializer
 
 import random
 import time
 
 @api_view(['GET'])
-def item_search(request):
-    start = time.time()
+def search(request):
 
     if request.method == 'GET':
 
@@ -29,7 +26,7 @@ def item_search(request):
             maxprice = request.query_params.get('maxprice')
             size = request.query_params.get('size')
             state = request.query_params.get('state')
-            # items = Item.objects.select_related('sell_username')
+
             items = Item.objects.all()
 
             if keyword:
@@ -40,11 +37,6 @@ def item_search(request):
                 데이터베이스에서 쿼리를 날렸을 때는 속도가 빨랐지만 장고에서 ORM으로 쿼리를 날릴 경우,
                 LIKE보다 살짝 느리다. 조금만 더 하면 성능 향상을 보일 듯하다. 
                 '''
-
-                # items = items.filter(
-                #     Q(title__icontains=keyword) |
-                #     Q(content__icontains=keyword)
-                # )
             
             if minprice and maxprice:
                 
@@ -86,11 +78,8 @@ def item_search(request):
             limit = int(request.query_params.get('limit'))
             items = items[offset:limit+offset]
             serializer = ItemSerializer(items, many=True)
-            print(items.query)
 
         except:
             return Response({"Error message":"Wrong offset or limit"}, status=status.HTTP_400_BAD_REQUEST)
 
-
-        print("time :", time.time() - start)
         return Response(serializer.data, status=status.HTTP_200_OK)
