+++ back/app/api/user/register/views_register.py
@@ -1,0 +1,25 @@
+from rest_framework.response import Response
+from rest_framework.decorators import api_view
+from rest_framework.authtoken.models import Token
+from rest_framework import status
+
+from app.models import User
+from app.middleware import validation_token
+from .serializer_register import RegisterationSerializer
+
+@api_view(['POST'])
+def registration(request):
+    
+    if request.method == 'POST':
+        serializer = RegisterationSerializer(data=request.data)
+        data = {}
+        if serializer.is_valid():
+            user = serializer.save()
+            data['response'] = "Registerd succesfully"
+            data['username'] = user.username
+            token = Token.objects.get(user=user).key
+            data['token'] = token 
+
+        else:
+            data = serializer.errors
+        return Response(data)
