+++ back/app/api/user/user_detail/serializer_user_detail.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from app.models import User
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ['id', 'username']
