+++ back/app/models.py
@@ -7,7 +7,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from rest_framework.authtoken.models import Token
 
-from .django_fulltext_search import SearchManager
+from .api.item.search.django_fulltext_search import SearchManager
 
 class User(AbstractUser):
     phone = models.CharField(max_length=15)
